Initial version of having a build-specific set of NPM dependencies
