+++ package.json
@@ -3,7 +3,10 @@
   "version": "1.0.0-dev",
   "description": "NodeJS/browser bindings to the aws-c-* libraries",
   "homepage": "https://github.com/awslabs/aws-crt-nodejs",
-  "repository": "github:awslabs/aws-crt-nodejs",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/awslabs/aws-crt-nodejs.git"
+  },
   "contributors": [
     "AWS Common Runtime Team <aws-sdk-common-runtime@amazon.com>"
   ],
@@ -46,7 +49,6 @@
     "@aws-sdk/util-utf8-browser": "^3.109.0",
     "@httptoolkit/websocket-stream": "^6.0.0",
     "axios": "^0.24.0",
-    "cmake-js": "^6.3.2",
     "crypto-js": "^4.0.0",
     "mqtt": "^4.3.7"
   }
