+++ scripts/build_dependencies/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "aws-crt-build",
+  "private": true,
+  "dependencies": {
+    "cmake-js": "^6.3.2"
+  },
+  "devDependencies": {}
+}
